requirements.txt
flask
flask-cors
requests
textblob
python-dotenv
schedule